% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.cdf_eval}
\alias{.cdf_eval}
\title{Get CDF function for a distribution}
\usage{
.cdf_eval(dist, params)
}
\arguments{
\item{dist}{Character name of distribution}

\item{params}{List of distribution parameters}
}
\value{
A function that evaluates the cdf
}
\description{
Get CDF function for a distribution
}
\keyword{internal}
