% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstats.R
\name{.compute_orderstat_mean}
\alias{.compute_orderstat_mean}
\title{Compute theoretical mean of order statistic}
\usage{
.compute_orderstat_mean(dist, params, n, r)
}
\description{
Compute theoretical mean of order statistic
}
\keyword{internal}
