% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderstats.R
\name{.compute_orderstat_var}
\alias{.compute_orderstat_var}
\title{Compute theoretical variance of order statistic}
\usage{
.compute_orderstat_var(dist, params, n, r)
}
\description{
Compute theoretical variance of order statistic
}
\keyword{internal}
