% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.default_params}
\alias{.default_params}
\title{Default parameters for distribution}
\usage{
.default_params(dist)
}
\arguments{
\item{dist}{Distribution name}
}
\value{
List of default parameters
}
\description{
Default parameters for distribution
}
\keyword{internal}
