% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.format_dist}
\alias{.format_dist}
\title{Format distribution for printing}
\usage{
.format_dist(dist, params)
}
\arguments{
\item{dist}{Distribution name}

\item{params}{Parameters}
}
\value{
Formatted string
}
\description{
Format distribution for printing
}
\keyword{internal}
