% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.pdf_eval}
\alias{.pdf_eval}
\title{Get PDF function for a distribution}
\usage{
.pdf_eval(dist, params)
}
\arguments{
\item{dist}{Character name of distribution}

\item{params}{List of distribution parameters}
}
\value{
A function that evaluates the pdf
}
\description{
Get PDF function for a distribution
}
\keyword{internal}
