% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plot_orderstat}
\alias{.plot_orderstat}
\title{Plot order statistic with theory vs simulation}
\usage{
.plot_orderstat(x)
}
\description{
Plot order statistic with theory vs simulation
}
\keyword{internal}
