% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{.plot_transform}
\alias{.plot_transform}
\title{Plot transformation with theory vs simulation}
\usage{
.plot_transform(x)
}
\description{
Plot transformation with theory vs simulation
}
\keyword{internal}
