% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.rng_eval}
\alias{.rng_eval}
\title{Get random number generator}
\usage{
.rng_eval(dist, params, n)
}
\arguments{
\item{dist}{Character name of distribution}

\item{params}{List of distribution parameters}

\item{n}{Number of observations}
}
\value{
Vector of random values
}
\description{
Get random number generator
}
\keyword{internal}
