% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.support_eval}
\alias{.support_eval}
\title{Get support of a distribution}
\usage{
.support_eval(dist, params)
}
\arguments{
\item{dist}{Character name of distribution}

\item{params}{List of distribution parameters}
}
\value{
List with lower and upper bounds
}
\description{
Get support of a distribution
}
\keyword{internal}
